requirements.txt
selenium==4.26.1
python-dotenv~=1.0.1
arsenic~=21.8
motor~=3.6.1